+++ addons/units/cfg/MSF_Uniforms.hpp
@@ -8,4 +8,5 @@ class Uniform_Base : ItemCore
 #include "uniforms\blufor\MSF_blk_Uniforms.hpp"
 #include "uniforms\blufor\MSF_mc_Uniforms.hpp"
 #include "uniforms\blufor\MSF_w_Uniforms.hpp"
+#include "uniforms\blufor\MSF_trop_Uniforms.hpp"
 #include "uniforms\opfor\MSF_arc_Uniforms.hpp"