+++ addons/units/cfg/uniforms/blufor/MSF_trop_Uniforms.hpp
@@ -1,0 +1,13 @@
+class MSF_NATO_Tanker_Trop : Uniform_Base
+{
+    author="Zeik_Tuvai";
+    scope = 2;
+    displayName = "Tanker Coveralls (Tropic)";
+    picture = "\A3\characters_f_exp\data\ui\icon_U_BT_Soldier_F_tna_ca.paa";
+    model = "\A3\Characters_F\Common\Suitpacks\suitpack_universal_F.p3d"; 
+    class ItemInfo : UniformItem { 
+        uniformClass = "MSF_Crewman_T"; 
+        containerClass = "Supply50"; 
+        mass = 50; 
+    };
+};