+++ addons/units/vehicles/blufor/mc/MSF_mc_Mine.hpp
@@ -6,6 +6,7 @@ class MSF_Mine :  MSF_Soldier_Base
 	scope = 2;
 	scopeCurator = 2;
 	canDeactivateMines = 1;
+	editorPreview = "z\msf\addons\units\ui\icon\blufor\mc\men\MSF_Mine.jpg";
 	icon = "iconManExplosive";
 	weapons[] = {
         "MSF_MX", "Binocular", "Throw", "Put"
@@ -18,7 +19,7 @@ class MSF_Mine :  MSF_Soldier_Base
 		"ACE_DefusalKit", "MineDetector", "ACE_Banana" 
     };
     linkedItems[] = {
-        "V_EOD_olive_F", "H_HelmetHBK_chops_F", "ItemMap", "ItemRadio", "ItemCompass", "ACE_Altimeter", "ItemGPS", "JAS_GPNVG18_blk"
+        "V_EOD_olive_F", "H_HelmetB_TI_arid_F", "ItemMap", "ItemRadio", "ItemCompass", "ACE_Altimeter", "ItemGPS", "JAS_GPNVG18_blk"
     };	
 	respawnWeapons[] = {
 		"MSF_MX", "Binocular", "Throw", "Put"
@@ -31,8 +32,8 @@ class MSF_Mine :  MSF_Soldier_Base
 		"ACE_DefusalKit", "MineDetector", "ACE_Banana"
 	};
 	respawnLinkedItems[] = {
-        "V_EOD_olive_F", "H_HelmetHBK_chops_F", "ItemMap", "ItemRadio", "ItemCompass", "ACE_Altimeter", "ItemGPS", "JAS_GPNVG18_blk"
+        "V_EOD_olive_F", "H_HelmetB_TI_arid_F", "ItemMap", "ItemRadio", "ItemCompass", "ACE_Altimeter", "ItemGPS", "JAS_GPNVG18_blk"
 	};
-    uniformClass = "U_B_T_Soldier_SL_F";
+    uniformClass = "U_B_CombatUniform_mcam";
     backpack = "MSF_mc_MineBackpack";
 };