+++ addons/units/vehicles/blufor/sf/MSF_sf_Diver.hpp
@@ -5,6 +5,7 @@ class MSF_SF_Diver :  MSF_Soldier_Base
 	editorSubcategory = "MSF_EdSubcat_Inf_Spec";
 	scope = 2;
 	scopeCurator = 2;
+	editorPreview = "\A3\EditorPreviews_F\Data\CfgVehicles\B_diver_F.jpg";
 	weapons[] = {
         "arifle_SDAR_F", "Binocular", "hgun_P07_snds_F", "Throw", "Put"
     };