+++ addons/units/vehicles/blufor/trop/MSF_trop_Rifleman.hpp
@@ -1,6 +1,7 @@
 class MSF_Rifleman_T :  MSF_Rifleman
 {
 	editorSubcategory = "MSF_EdSubcat_Inf_Trop";
+    editorPreview = "z\msf\addons\units\ui\icon\blufor\trop\men\MSF_Rifleman_T.jpg";
     #include "mx.inc"
     linkedItems[] = {
         "V_PlateCarrier2_tna_F", "H_HelmetB_Enh_tna_F", "JAS_GPNVG18_blk",